资源管理 优化代码 防止在控制器中编写数据库操作代码

+++ src/main/java/model/com/iisquare/jees/oa/service/ResourceService.java
@@ -1,6 +1,5 @@
 package com.iisquare.jees.oa.service;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -10,10 +9,10 @@ import org.springframework.stereotype.Service;
 import com.iisquare.jees.framework.model.ServiceBase;
 import com.iisquare.jees.framework.util.DPUtil;
 import com.iisquare.jees.framework.util.ServiceUtil;
-import com.iisquare.jees.framework.util.SqlUtil;
 import com.iisquare.jees.oa.dao.MemberDao;
 import com.iisquare.jees.oa.dao.ResourceDao;
 import com.iisquare.jees.oa.domain.Resource;
+import com.iisquare.jees.oa.domain.Role;
 
 @Service
 public class ResourceService extends ServiceBase {
@@ -25,42 +24,18 @@ public class ResourceService extends ServiceBase {
 	
 	public ResourceService() {}
 	
-	public Map<Object, Object> search(Map<String, String> map, String orderBy, int page, int pageSize) {
-		StringBuilder sb = new StringBuilder("select * from ")
-			.append(resourceDao.tableName()).append(" where 1 = 1");
-		Map<String, Object> paramMap = new HashMap<String, Object>();
-		Object name = map.get("name");
-		if(!DPUtil.empty(name)) {
-			sb.append(" and name like :name");
-			paramMap.put("name", DPUtil.stringConcat("%", name, "%"));
+	public List<Map<String, Object>> getList(String columns, boolean bRefer, String orderBy, int page, int pageSize) {
+		String append = null;
+		if(!DPUtil.empty(orderBy)) append = DPUtil.stringConcat(" order by ", orderBy);
+		List<Map<String, Object>> list;
+		if(bRefer) {
+			list = resourceDao.getPage(columns, new String[]{"refer_id"}, new Object[]{0}, new String[]{">"}, append, page, pageSize);
+		} else {
+			list = resourceDao.getPage(columns, null, null, append, page, pageSize);
 		}
-		Object module = map.get("module");
-		if(!DPUtil.empty(module)) {
-			sb.append(" and module = :module");
-			paramMap.put("module", module);
-		}
-		Object controller = map.get("controller");
-		if(!DPUtil.empty(controller)) {
-			sb.append(" and controller = :controller");
-			paramMap.put("controller", controller);
-		}
-		Object action = map.get("action");
-		if(!DPUtil.empty(action)) {
-			sb.append(" and action = :action");
-			paramMap.put("action", action);
-		}
-		Object referId = map.get("referId");
-		if(null != referId && !"".equals(referId)) {
-			sb.append(" and refer_id = :referId");
-			paramMap.put("referId", DPUtil.parseInt(referId));
-		}
-		if(!DPUtil.empty(orderBy)) sb.append(" order by ").append(orderBy);
-		String sql = sb.toString();
-		int total = resourceDao.getCount(sql, paramMap, true);
-		sql = DPUtil.stringConcat(sql, SqlUtil.buildLimit(page, pageSize));
-		List<Map<String, Object>> rows = resourceDao.npJdbcTemplate().queryForList(sql, paramMap);
-		rows = ServiceUtil.fillRelations(rows, memberDao, new String[]{"create_id", "update_id"}, new String[]{"serial", "name"}, null);
-		return DPUtil.buildMap(new String[]{"total", "rows"}, new Object[]{total, rows});
+		list = ServiceUtil.fillProperties(list, new Role(), new String[]{"status"}, new String[]{"statusText"}, true);
+		list = ServiceUtil.fillRelations(list, memberDao, new String[]{"create_id", "update_id"}, new String[]{"serial", "name"}, null);
+		return list;
 	}
 	
 	public Resource getById(Object id) {
@@ -90,6 +65,8 @@ public class ResourceService extends ServiceBase {
 	}
 	
 	public int delete(Object... ids) {
+		int count = resourceDao.getCount(new String[]{"parent_id"}, new Object[]{ids}, new String[]{"in"}, null);
+		if(count > 0) return -1;
 		return resourceDao.deleteByIds(ids);
 	}
 }
